Revert move to RenderEffect commit_hash:51a8a6f7d66ef6bb7679743bd8513a6607f061f2

+++ client/android/div/src/main/java/com/yandex/div/core/view2/divs/BaseDivViewExtensions.kt
@@ -863,7 +863,7 @@ internal fun View.applyBitmapFilters(
             when (filter) {
                 is DivFilter.Blur -> {
                     val radius = filter.value.radius.evaluate(resolver).toIntSafely().dpToPx(resources.displayMetrics)
-                    result = bitmapEffectHelper.blurBitmap(result, radius.toFloat())
+                    result = bitmapEffectHelper.blurBitmap(result, radius)
                 }
 
                 is DivFilter.RtlMirror -> if (isLayoutRtl()) {
